2 buttons and adjustment
added two buttons for linked lists and trees. moved array class code to a new .js file.
